Add files via upload
The problem of covering random points in the plane with disks has applications in operations research, as well as being of theoretical interest. In this document we show that the computational complexity may be reduced by classifying points and into several classes (non-covered, single-covered, collateral, and indeterminate) and sets into several classes (non-covering,single-covering, collateral, redundant, and indeterminate). Sets that are single-covering are necessary for the cover, while non-covering, collateral, and redundant sets can be excluded.  An optimal cover can be found by applying linear programming to just the indeterminate points and sets, which may be much fewer than the original points and sets given. In addition, indeterminate points and sets may be divided into separate “islands” that can be solved separately.Hence the actual complexity is determined by the number of points and sets in the largest island.We give several heatmaps to show how the proportion of points and sets of various types depend on two basic scale-invariant parameters related to point and set density. For a few parameters,we compare these results to some theoretical calculations.